Implement changes in firmware v4.2
* Source indexes now uses 16bits
* All gain settings has been changed to have a common gain for left
and right channel.

+++ qdownstreamkeysettings.h
@@ -45,8 +45,8 @@ struct QDownstreamKeySettings
     bool m_tie;
     quint8 m_frames;
     quint8 m_frameCount;
-    quint8 m_fillSource;
-    quint8 m_keySource;
+    quint16 m_fillSource;
+    quint16 m_keySource;
     bool m_invertKey;
     bool m_preMultiplied;
     float m_clip;
